+++ packages/plugin-svelte-query/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @kubb/plugin-svelte-query
 
+## 4.20.0
+
+### Patch Changes
+
+- [#2387](https://github.com/kubb-labs/kubb/pull/2387) [`d3acf9e`](https://github.com/kubb-labs/kubb/commit/d3acf9eb2b018595fadcc06380ef8419d8bbea8f) Thanks [@stijnvanhulle](https://github.com/stijnvanhulle)! - Update fabric
+
+- Updated dependencies [[`d3acf9e`](https://github.com/kubb-labs/kubb/commit/d3acf9eb2b018595fadcc06380ef8419d8bbea8f)]:
+  - @kubb/plugin-client@4.20.0
+  - @kubb/plugin-oas@4.20.0
+  - @kubb/plugin-zod@4.20.0
+  - @kubb/plugin-ts@4.20.0
+  - @kubb/core@4.20.0
+  - @kubb/oas@4.20.0
+
 ## 4.19.2
 
 ### Patch Changes